Refactor Code Template for improved readability

+++ Code Template.cpp
@@ -44,7 +44,7 @@ int main() {
     }
     Longgggg;
 
-    ll t = 1;
+    int t = 1;
     // cin >> t;
     while (t--) solve();
     return 0;
